More specific answer
The given answer failed to account for qty values above 20. In the case of the "banana" item, it returned an freeQty value of 6.  In the exercise description it states "Purchases of 5 or more receive 1 free.  Purchases of 10 or more receive 3 free". The original code added +3 for each 10 instead of +3 if qty is >= 10.  Great course otherwise :) .