changes for the file to work with gulp 4 and node 13
There is a problem when we try to run gulp 3 with node 12. So to fix this I had to change my vertions. I've got gulp 4 and node 13 wich work fine. The changes here are to adapt the gulp file to gulp 4 wich has some differences in writing [ https://stackoverflow.com/a/54939126/9208292 ].